% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{connect_sqlite_db}
\alias{connect_sqlite_db}
\title{Connect to the Pediatric Drug Safety database}
\usage{
connect_sqlite_db(dbname = "./effect_peds_19q2_v0.3_20211119.sqlite")
}
\arguments{
\item{dbname}{The path to the sqlite file}
}
\value{
SQLite connection, invisibly
}
\description{
This function makes a sqlite connection from the downloaded database.
}
