% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{disconnect_sqlite_db}
\alias{disconnect_sqlite_db}
\title{Disconnect from the Pediatric Drug Safety database}
\usage{
disconnect_sqlite_db(con)
}
\arguments{
\item{con}{The sqlite connection}
}
\value{
TRUE, invisibly
}
\description{
Disconnect the sqlite database connection.
}
\examples{
if(interactive()){
download_sqlite_db()
con <- connect_sqlite_db()
disconnect_sqlite_db(con)
}
}
