% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_sqlite_db}
\alias{download_sqlite_db}
\title{Download the Pediatric Drug Safety database}
\usage{
download_sqlite_db(
  url = "https://pds-database.s3.amazonaws.com/effect_peds_19q2_v0.3_20211119.sqlite",
  path = "./",
  destfile = paste0(path, "effect_peds_19q2_v0.3_20211119.sqlite"),
  method = "auto",
  quiet = TRUE,
  timeout = 360
)
}
\arguments{
\item{url}{The url of the sqlite database}

\item{path}{The path where to dump the sqlite database.}

\item{destfile}{The path and filename where to dump the sqlite database. See \code{download.file}}

\item{method}{The method to download the sqlite database. See \code{download.file}}

\item{quiet}{Whether to download quietly. See \code{download.file}}

\item{timeout}{Extended download session for downloading tihs file. Default is 360 seconds.}
}
\value{
TRUE, invisibly
}
\description{
Thus function downloads the database published in Giangreco et al. 2022.
}
