% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_db_path}
\alias{get_db_path}
\title{Return database cache}
\usage{
get_db_path()
}
\value{
list
}
\description{
This function returns the URL, sqlite database file, and cache names to be used for downloading the database to your machine.
}
\examples{
get_db_path()

}
