% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{connect_sqlite_db}
\alias{connect_sqlite_db}
\title{Connect to the Pediatric Drug Safety database}
\usage{
connect_sqlite_db()
}
\value{
SQLite connection
}
\description{
Establish a sqlite connection from the downloaded database.
}
\examples{
if(FALSE){
download_sqlite_db()
con <- connect_sqlite_db()
disconnect_sqlite_db(con)
}
}
