% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_sqlite_db}
\alias{download_sqlite_db}
\title{Download the Pediatric Drug Safety database}
\usage{
download_sqlite_db(
  method = "auto",
  quiet = FALSE,
  timeout = 1000,
  force = FALSE
)
}
\arguments{
\item{method}{The method to download the sqlite database. See \code{download.file}}

\item{quiet}{Whether to download quietly. See \code{download.file}}

\item{timeout}{Extended download session for downloading this file. Default is 1000 seconds.}

\item{force}{Whether to force the download of the database. Defaults to FALSE.}
}
\value{
TRUE, invisibly
}
\description{
Download the database published in Giangreco et al. 2022. Warning, the size of the uncompressed 'sqlite' file is close to 900 MB. Use wit caution
}
\examples{
if(interactive()){
download_sqlite_db()
}
}
