% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_sqlite_db}
\alias{download_sqlite_db}
\title{Download the Pediatric Drug Safety database}
\usage{
download_sqlite_db(
  method = "auto",
  quiet = FALSE,
  timeout = 360,
  force = FALSE
)
}
\arguments{
\item{method}{The method to download the sqlite database. See \code{download.file}}

\item{quiet}{Whether to download quietly. See \code{download.file}}

\item{timeout}{Extended download session for downloading tihs file. Default is 360 seconds.}

\item{force}{Whether to force the download of the database. Defaults to FALSE.}
}
\value{
TRUE, invisibly
}
\description{
Download the database published in Giangreco et al. 2022.
}
\examples{
\dontrun{
download_sqlite_db()
}
}
