% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_sqlite_db}
\alias{download_sqlite_db}
\title{Download the Pediatric Drug Safety database}
\usage{
download_sqlite_db(
  method = "auto",
  quiet = FALSE,
  timeout = 1000,
  force = FALSE
)
}
\arguments{
\item{method}{The method to download the sqlite database. See \code{download.file}}

\item{quiet}{Whether to download quietly. See \code{download.file}}

\item{timeout}{Extended download session for downloading this file. Default is 1000 seconds.}

\item{force}{Whether to force the download of the database. Defaults to FALSE. Needs to be TRUE for database to download. The function will prompt for confirmation.}
}
\value{
TRUE, invisibly
}
\description{
Download the database published in Giangreco et al. 2022. This function will prompt to download the database, so the cache directory will be identified and the database will be downloaded to it only after consent. Warning, the size of the uncompressed 'sqlite' file is close to 0.9GB or 900 MB. Use with caution.
}
\examples{
if(FALSE){
download_sqlite_db() #set force=TRUE if desired to download 0.9GB file to machine
}
}
