% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_db_path}
\alias{get_db_path}
\title{Return database cache}
\usage{
get_db_path()
}
\value{
list
}
\description{
This function identifies and returns the cache location for the database on your machine
}
\examples{
get_db_path()

}
